requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.2
kubernetes==26.1.0
openai==0.27.4
python-dotenv